+++ src/api/java/com/pg85/otg/forge/ForgeWorld.java
@@ -1,7 +1,0 @@
-package com.pg85.otg.forge;
-
-import net.minecraft.world.World;
-
-public abstract class ForgeWorld {
-    public World world;
-}