+++ src/api/java/com/pg85/otg/forge/generator/ForgeChunkBuffer.java
@@ -1,8 +1,0 @@
-package com.pg85.otg.forge.generator;
-
-import com.pg85.otg.generator.ChunkBuffer;
-import net.minecraft.world.chunk.ChunkPrimer;
-
-abstract class ForgeChunkBuffer implements ChunkBuffer {
-    private ChunkPrimer chunkPrimer;
-}