+++ src/api/java/com/pg85/otg/generator/ChunkBuffer.java
@@ -1,7 +1,0 @@
-package com.pg85.otg.generator;
-
-import com.pg85.otg.util.ChunkCoordinate;
-
-public interface ChunkBuffer {
-    ChunkCoordinate getChunkCoordinate();
-}