+++ src/api/java/com/pg85/otg/generator/ChunkProviderOTG.java
@@ -1,8 +1,0 @@
-package com.pg85.otg.generator;
-
-import com.pg85.otg.forge.ForgeWorld;
-
-public abstract class ChunkProviderOTG {
-    private ForgeWorld localWorld;
-    protected abstract void generateTerrain(ChunkBuffer chunkBuffer);
-}