+++ src/main/scala/farseek/world/gen/OtgHook.scala
@@ -1,6 +1,7 @@
-package com.pg85.otg.forge.generator
+package farseek.world.gen
 
 import com.pg85.otg.forge.ForgeWorld
+import com.pg85.otg.forge.generator._
 import com.pg85.otg.generator._
 import farseek.util.Reflection._
 import net.minecraft.world.WorldServer
@@ -12,7 +13,6 @@ object OtgHook {
   private val otgGenerateTerrain = classOf[ChunkProviderOTG ].getDeclaredMethod("generateTerrain", classOf[ChunkBuffer]).accessible
   private val otgWorld           = classOf[ChunkProviderOTG ].getDeclaredField("localWorld").accessible
   private val otgGenerator       = classOf[OTGChunkGenerator].getDeclaredField("generator").accessible
-  private val otgChunkPrimer     = classOf[ForgeChunkBuffer ].getDeclaredField("chunkPrimer").accessible // package-private class
 
   def generateTerrain(generator: ChunkProviderOTG, chunkBuffer: ChunkBuffer): Unit = {
     otgGenerateTerrain(generator, chunkBuffer)
@@ -20,7 +20,8 @@ object OtgHook {
     val worldChunkGenerator = world.getChunkProvider.chunkGenerator.asInstanceOf[OTGChunkGenerator]
     if(generator eq otgGenerator(worldChunkGenerator)) {
       val chunkCoords = chunkBuffer.getChunkCoordinate
-      EVENT_BUS.post(new ReplaceBiomeBlocks(worldChunkGenerator, chunkCoords.getChunkX, chunkCoords.getChunkZ, otgChunkPrimer(chunkBuffer), world))
+      EVENT_BUS.post(new ReplaceBiomeBlocks(worldChunkGenerator, chunkCoords.getChunkX, chunkCoords.getChunkZ,
+          chunkBuffer.getClass.getDeclaredField("chunkPrimer").accessible(chunkBuffer), world))
     }
   }
 }